+++ lib/kitabu/extensions/rouge.rb
@@ -5,6 +5,8 @@ module Rouge
     module Redcarpet
       def rouge_formatter(lexer)
         options = lexer.respond_to?(:options) ? lexer.options : {}
+        options = options.keys.map(&:to_sym).zip(options.values).to_h
+        options[:start_line] = options.fetch(:start_line, 1).to_i
 
         Formatters::HTMLLegacy.new(
           {css_class: "highlight #{lexer.tag}"}.merge(options)